requirements.txt
ultralytics
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
pillow